+++ .pre-commit-config.yaml
@@ -1,0 +1,8 @@
+repos:
+  - repo: git://github.com/dnephin/pre-commit-golang
+    rev: v0.3.5
+    hooks:
+    - id: go-fmt
+    - id: go-lint
+    - id: validate-toml
+    - id: no-go-testing
