+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/wangkuiyi/gotorch
+
+go 1.13
